requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0


# Authentication & Security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data validation
email-validator==2.1.1

# Environment variables
python-dotenv==1.0.1

# Utilities
requests==2.31.0
